+++ apps/daemon/src/artifact-manifest.ts
