+++ apps/daemon/src/claude-stream.ts
