+++ apps/daemon/src/cli.ts
