+++ apps/daemon/src/copilot-stream.ts
