+++ apps/daemon/src/db.ts
