+++ apps/daemon/src/document-preview.ts
