+++ apps/daemon/src/json-event-stream.ts
