+++ apps/daemon/src/lint-artifact.ts
