+++ apps/daemon/src/projects.ts
