+++ apps/daemon/src/skills.ts
