+++ apps/web/src/providers/daemon.ts
@@ -168,7 +168,7 @@ export async function streamViaDaemon({
   }
 }
 
-// Translate a raw `agent` SSE payload (what apps/daemon/claude-stream.js emits)
+// Translate a raw `agent` SSE payload (what apps/daemon/src/claude-stream.ts emits)
 // into the UI's AgentEvent union. Keep this liberal — unknown types just
 // return null so the UI ignores them instead of rendering garbage.
 function translateAgentEvent(data: DaemonAgentPayload): AgentEvent | null {
