+++ e2e/tests/structured-streams.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { createClaudeStreamHandler } from '../../apps/daemon/claude-stream.js';
-import { createCopilotStreamHandler } from '../../apps/daemon/copilot-stream.js';
+import { createClaudeStreamHandler } from '../../apps/daemon/src/claude-stream.js';
+import { createCopilotStreamHandler } from '../../apps/daemon/src/copilot-stream.js';
 
 describe('structured agent stream fixtures', () => {
   it('emits TodoWrite tool_use from Claude Code stream JSON', () => {
